# frontend/src/api/generated/current-openapi.yaml
openapi: 3.1.0
info:
  title: ShuttleCube Internal Management API
  version: 0.1.0
  description: |
    羽毛球培训与场地经营管理系统的同源内部接口契约。
    该接口仅供已登录管理界面使用，不是外部公共 API。
    `/agents`、`/workflows`、`/agent-runs` 和 `/approval-requests` 为未来 Agent 契约草案，
    当前版本不得实现、生成客户端调用或作为验收范围。
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /classes/{class_id}:
    get:
      tags:
        - Classes
      summary: 查询固定班课程、报名与收费详情
      operationId: getClass
      parameters:
        - name: class_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 固定班详情，报名收费按实际资金事实汇总
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FixedClassDetail"
        "404":
          description: 固定班不存在
...
  /events:
    post:
      tags:
        - Events
      summary: 创建临时活动
      operationId: createEvent
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventWrite"
      responses:
        "201":
          description: 活动已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessCreated"
        "409":
          $ref: "#/components/responses/ScheduleConflict"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /coach-fees:
    get:
      tags:
        - Payroll
      summary: 查询教练费用
      operationId: listCoachFees
      parameters:
        - name: coach_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: status
          in: query
          schema:
            type: string
            enum:
              - pending
              - settled
              - void
      responses:
        "200":
          description: 教练费用明细和汇总
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachFeeSummary"
  /coach-fees/{fee_id}:
    get:
      tags:
        - Payroll
      summary: 查询单笔教练费用
      operationId: getCoachFee
      parameters:
        - name: fee_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 教练费用详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachFee"
        "404":
          description: 教练费用不存在
    patch:
      tags:
        - Payroll
      summary: 有原因地调整待结教练费用
      operationId: adjustCoachFee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - adjustment_amount
                - reason
                - version
              properties:
                adjustment_amount:
                  $ref: "#/components/schemas/Money"
                reason:
                  type: string
                  minLength: 1
                version:
                  type: integer
                  minimum: 1
      responses:
        "200":
          description: 教练费用已调整
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachFee"
        "409":
          $ref: "#/components/responses/ConcurrentChange"
  /payroll-settlements:
    get:
      tags:
        - Payroll
      summary: 查询教练结算历史
      operationId: listPayrollSettlements
      parameters:
        - name: coach_id
          in: query
          schema:
            type: string
            format: uuid
        - name: period_month
          in: query
          description: 所查自然月的任意日期，服务端归一为月初
          schema:
            type: string
            format: date
      responses:
        "200":
          description: 结算历史
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PayrollSettlementResult"
    post:
      tags:
        - Payroll
      summary: 结算教练费用并生成工资支出
      operationId: createPayrollSettlement
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayrollSettlementWrite"
      responses:
        "201":
          description: 结算和工资支出已原子创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayrollSettlementResult"
        "409":
          $ref: "#/components/responses/ConcurrentChange"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /payroll-settlements/{settlement_id}/void:
    post:
      tags:
        - Payroll
      summary: 冲正错误结算并恢复待结费用
      operationId: voidPayrollSettlement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoidWrite"
      responses:
        "200":
          description: 结算已冲正
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayrollSettlementResult"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: shuttlecube_session
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 200
...
  responses:
...
    ValidationError:
      description: 输入或业务规则校验失败
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ValidationProblem"
    ConcurrentChange:
      description: 记录版本已变化或目标事实已被其他操作处理
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ScheduleConflict:
      description: 一个或多个资源发生排期冲突
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ScheduleConflictProblem"
...
  schemas:
    IdentifierVersion:
      type: object
      required:
        - id
        - version
      properties:
        id:
          type: string
          format: uuid
        version:
          type: integer
          minimum: 1
    Money:
      type: string
      pattern: ^\d+(\.\d{1,2})?$
      examples:
        - "120.00"
    Problem:
      type: object
      additionalProperties: true
      required:
        - type
        - title
        - status
        - detail
        - instance
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
    ValidationProblem:
      allOf:
        - $ref: "#/components/schemas/Problem"
        - type: object
          properties:
            errors:
              type: array
              items:
                type: object
                required:
                  - field
                  - message
                properties:
                  field:
                    type: string
                  message:
                    type: string
    ScheduleConflictProblem:
      allOf:
        - $ref: "#/components/schemas/Problem"
        - type: object
          required:
            - conflicts
          properties:
            conflicts:
              type: array
              items:
                $ref: "#/components/schemas/ScheduleConflict"
...
    ResourceRef:
      type: object
      required:
        - resource_type
        - resource_id
      properties:
        resource_type:
          type: string
          enum:
            - court
            - coach
            - student
        resource_id:
          type: string
          format: uuid
        display_name:
          type: string
...
    ScheduleConflict:
      type: object
      required:
        - resource
        - starts_at
        - ends_at
        - conflicting_schedule_id
      properties:
        resource:
          $ref: "#/components/schemas/ResourceRef"
        starts_at:
          type: string
          format: date-time
        ends_at:
          type: string
          format: date-time
        conflicting_schedule_id:
          type: string
          format: uuid
        reason:
          type: string
...
    EnrollmentCashSummary:
      type: object
      required:
        - receivable_id
        - actual_amount
        - received_amount
        - refunded_amount
        - net_received
        - outstanding_amount
        - payment_status
      properties:
        receivable_id:
          type: string
          format: uuid
        actual_amount:
          $ref: "#/components/schemas/Money"
        received_amount:
          $ref: "#/components/schemas/Money"
        refunded_amount:
          $ref: "#/components/schemas/Money"
        net_received:
          $ref: "#/components/schemas/Money"
        outstanding_amount:
          $ref: "#/components/schemas/Money"
        payment_status:
          type: string
          enum:
            - unpaid
            - partial
            - paid
            - partially_refunded
            - refunded
    FixedClassCashSummary:
      type: object
      required:
        - actual_amount
        - received_amount
        - refunded_amount
        - net_received
        - outstanding_amount
      properties:
        actual_amount:
          $ref: "#/components/schemas/Money"
        received_amount:
          $ref: "#/components/schemas/Money"
        refunded_amount:
          $ref: "#/components/schemas/Money"
        net_received:
          $ref: "#/components/schemas/Money"
        outstanding_amount:
          $ref: "#/components/schemas/Money"
    FixedClassEnrollmentDetail:
      type: object
      required:
        - id
        - student_id
        - student_name
        - purchased_units
        - actual_receivable
        - status
        - finance
      properties:
        id:
          type: string
          format: uuid
        student_id:
          type: string
          format: uuid
        student_name:
          type: string
        purchased_units:
          type: integer
          minimum: 0
        remaining_units:
          type: integer
          minimum: 0
        unit_price:
          $ref: "#/components/schemas/Money"
        actual_receivable:
          $ref: "#/components/schemas/Money"
        status:
          type: string
        acquisition_type:
          type: string
          enum:
            - purchase
            - transfer
        finance:
          anyOf:
            - $ref: "#/components/schemas/EnrollmentCashSummary"
            - type: "null"
    FixedClassDetail:
      type: object
      required:
        - id
        - name
        - status
        - version
        - session_count
        - capacity
        - coach_fee_per_session
        - finance
        - sessions
        - enrollments
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
        version:
          type: integer
          minimum: 1
        session_count:
          type: integer
          minimum: 1
        capacity:
          type: integer
          minimum: 1
        coach_fee_per_session:
          $ref: "#/components/schemas/Money"
        finance:
          $ref: "#/components/schemas/FixedClassCashSummary"
        sessions:
          type: array
          items:
            type: object
            required:
              - id
              - sequence_number
              - scheduled_start
              - scheduled_end
              - status
              - attendance_finalized_at
              - attendance
              - coach_fee
            properties:
              id:
                type: string
                format: uuid
              version:
                type: integer
                minimum: 1
              sequence_number:
                type: integer
                minimum: 1
              scheduled_start:
                type: string
                format: date-time
              scheduled_end:
                type: string
                format: date-time
              status:
                type: string
              replacement_for_session_id:
                type:
                  - string
                  - "null"
                format: uuid
              replacement_for_sequence:
                type:
                  - integer
                  - "null"
                minimum: 1
              replacement_decision:
                type:
                  - string
                  - "null"
                enum:
                  - pending
                  - scheduled
                  - waived
                  - null
              attendance_finalized_at:
                type:
                  - string
                  - "null"
                format: date-time
              attendance:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceResult"
              coach_fee:
                anyOf:
                  - $ref: "#/components/schemas/GeneratedCoachFee"
                  - type: "null"
        enrollments:
          type: array
          items:
            $ref: "#/components/schemas/FixedClassEnrollmentDetail"
...
    AttendanceResult:
      type: object
      required:
        - id
        - student_id
        - student_name
        - status
        - deduct_units
        - grants_makeup
        - decision_note
      properties:
        id:
          type: string
          format: uuid
        student_id:
          type: string
          format: uuid
        student_name:
          type: string
        status:
          type: string
          enum:
            - present
            - leave
            - absent
            - makeup_present
            - unprocessed
        deduct_units:
          type: integer
          minimum: 0
          maximum: 1
        grants_makeup:
          type: boolean
        decision_note:
          type:
            - string
            - "null"
...
    GeneratedCoachFee:
      type: object
      required:
        - id
        - amount
        - status
      properties:
        id:
          type: string
          format: uuid
        base_amount:
          $ref: "#/components/schemas/Money"
        adjustment_amount:
          $ref: "#/components/schemas/Money"
        amount:
          $ref: "#/components/schemas/Money"
        status:
          type: string
          enum:
            - pending
            - settled
            - void
        settlement_id:
          type:
            - string
            - "null"
          format: uuid
...
    EventWrite:
      type: object
      required:
        - event_type
        - name
        - starts_at
        - ends_at
        - court_ids
        - actual_receivable
      properties:
        event_type:
          type: string
        name:
          type: string
        starts_at:
          type: string
          format: date-time
        ends_at:
          type: string
          format: date-time
        court_ids:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
        coach_ids:
          type: array
          items:
            type: string
            format: uuid
        participant_ids:
          type: array
          items:
            type: string
            format: uuid
        track_participants:
          type: boolean
          default: false
        requires_attendance:
          type: boolean
          default: false
        actual_receivable:
          $ref: "#/components/schemas/Money"
        coach_fee:
          $ref: "#/components/schemas/Money"
        notes:
          type:
            - string
            - "null"
        warning_acknowledgements:
          type: array
          items:
            type: string
            enum:
              - past_time
              - outside_business_hours
...
    BusinessCreated:
      type: object
      required:
        - id
        - schedule_id
        - receivable_id
        - version
      properties:
        id:
          type: string
          format: uuid
        schedule_id:
          type: string
          format: uuid
        receivable_id:
          type: string
          format: uuid
        version:
          type: integer
...
    VoidWrite:
      type: object
      required:
        - reason
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 500
...
    CoachFeeSummary:
      type: object
      required:
        - coach_id
        - calculated_amount
        - items
      properties:
        coach_id:
          type:
            - string
            - "null"
          format: uuid
        calculated_amount:
          $ref: "#/components/schemas/Money"
        items:
          type: array
          items:
            $ref: "#/components/schemas/CoachFee"
    CoachFee:
      allOf:
        - $ref: "#/components/schemas/IdentifierVersion"
        - type: object
          required:
            - coach_id
            - coach_name
            - source_type
            - source_id
            - occurred_at
            - base_amount
            - adjustment_amount
            - amount
            - status
            - business_name
            - business_path
          properties:
            coach_id:
              type: string
              format: uuid
            coach_name:
              type: string
            source_type:
              type: string
              enum:
                - class_session
                - private_lesson
                - event
            source_id:
              type: string
              format: uuid
            occurred_at:
              type: string
              format: date-time
            base_amount:
              $ref: "#/components/schemas/Money"
            adjustment_amount:
              $ref: "#/components/schemas/Money"
            amount:
              $ref: "#/components/schemas/Money"
            adjustment_reason:
              type:
                - string
                - "null"
            status:
              type: string
              enum:
                - pending
                - settled
                - void
            settlement_id:
              type:
                - string
                - "null"
              format: uuid
            business_name:
              type: string
            business_path:
              type: string
            source_status:
              type:
                - string
                - "null"
    PayrollSettlementWrite:
      type: object
      required:
        - coach_id
        - period_month
        - actual_amount
        - paid_at
      properties:
        coach_id:
          type: string
          format: uuid
        period_month:
          type: string
          format: date
          description: 自然月任意日期；服务端自动纳入该月全部待结费用
        actual_amount:
          $ref: "#/components/schemas/Money"
        adjustment_reason:
          type:
            - string
            - "null"
        paid_at:
          type: string
          format: date-time
    PayrollSettlementResult:
      type: object
      required:
        - settlement_id
        - expense_id
        - settled_fee_ids
        - actual_amount
        - version
      properties:
        settlement_id:
          type: string
          format: uuid
        expense_id:
          type: string
          format: uuid
        settled_fee_ids:
          type: array
          items:
            type: string
            format: uuid
        actual_amount:
          $ref: "#/components/schemas/Money"
        version:
          type: integer
...